implement optimistic concurrency control
How can we prevent data race?

The simplest and cleanest approach in this case is to use a form of
optimistic locking based on the version number in our movie record.

This means that the first update request that reaches the database
will succeed, and whoever is making the second update will receive an
error message instead of having their change applied.

To make this work, we change the SQL statement for
updating a movie.