add a new middleware to recover the panic
At the moment any panics in our API handlers will be recovered
automatically by Go's http.Server. This will unwind the stack for the
affected goroutine (calling any deferred functions along the way),
close the underlying HTTP connection, and log an error message and
stack trace.

This behavior is OK, but it would be better for the client if we
could also a "500 Internal Server Error" response to explain that
something has gone wrong — rather than just closing the HTTP
connection with no context.