recover panic in background goroutines
It's important to bear in mind that any panic which happens in
background goroutine will not be automatically recovered by
`recoverPanic()` middleware or Go's `http.Server`, and will cause our
whole app to terminate.

So we need to make sure that any panic in background goroutine is
manually recovered, using a similar pattern to the one in
`recoverPanic()` middleware.