vendoring dependencies
Mitigate the risk of a disappearing dependency.

Go's module mirror functionality is great. But it isn't a silver
bullet for all developers and all projects.

So, for these reasons, it can still be sensible to vendor project
dependencies using the `go mod vendor` command. Vendoring
dependencies in this way basically stores a complete copy of the
source code for third-party packages in a `vendor` folder in our
project.

Changes:
- Adapt Makefile to include a new `vendor` rule.