deps: add realip package
Although we could write the logic to check for the presence of an
`X-Forwarded-For` header and handle it ourselves, we will use the
realip package to help with this. This package retrieves the client
IP address from any `X-Forwarded-For` or `X-Real-IP` headers, falling
back to use `r.RemoteAddr` if neither of them are present.