deps: add pascaldekloe/jwt package
There are a few different packages available which make working with
JWTs relatively simple in Go.
In most cases the github.com/pascaldekloe/jwt package is a good
choice — it has a clear and simple API, and is designed to avoid a
couple of the major JWT security vulnerabilities by default.