JWT based authentication
When using JWTs, you need to consider the choice of signing algorithm.
We'll use a (simpler and faster) symmetric-key algorithm HMAC-SHA256
with a random secret key for our API.

Changes:
* To get authentication with JWTs working, you'll first need to add
a secret key for signing the JWTs to your `.envrc` file.
* Then update Makefile to pass in the secret key as a command-line
flag when starting the application.
* Edit main function so that it parses the JWT secret from the
command-line-flag into the `config` struct.
* Change the `createAuthenticationTokenHandler()` so that it
generates and sends a JWT instead of a stateful token.
* Update the `authenticate()` middleware to accept JWTs in an
`Authorization: Bearer <jwt>` header, verify the JWT, and extract the
user ID from the `subject` claim.
* Create a new `UserModel.Get()` method to retrieve the user details
from the database based on their ID.