test: integration testing
Running end-to-end tests with mocked dependencies is a good thing to
do, but we could improve confidence in the app even more if we also
verify that the production PostgreSQL database models are working as
expected.

To do this we can run integration tests against a test version of
PostgreSQL database, which mimics the production database but exists
for testing purposes only.

Setup an integration test to ensure that the
`data.UserModel.GetByEmail()` method is working correctly.

Changes:
* Test database setup and teardown
Create the test version of our PostgreSQL database -- execute SQL
statements to create a new `test_skel` database.
Then, make two SQL scripts: setup script and teardown script.
* Create a new file to hold some helper functions.
Add to it a `newTestDB()` helper function.
* Test the `UserModel.GetByEmail` method.
Create a new file for this test `users_test.go`.